refactor: extract account restoration manager
