+++ src/App/bootstrap/app-stage2.tsx
@@ -9,15 +9,11 @@ import ConnectionErrorListener from "~Toasts/components/ConnectionErrorListener"
 import NotificationContainer from "~Toasts/components/NotificationContainer"
 import StellarUriHandler from "~TransactionRequest/components/StellarUriHandler"
 import WalletConnectHandler from "~WalletConnect/components/WalletConnectHandler"
-import { Account, AccountsContext } from "../contexts/accounts"
-import { SettingsContext } from "../contexts/settings"
 import AllAccountsPage from "../components/AccountListView"
 import AndroidBackButton from "../components/AndroidBackButton"
 import DesktopNotifications from "../components/DesktopNotifications"
+import LastOpenedAccountManager from "../components/LastOpenedAccountManager"
 import LinkHandler from "../components/LinkHandler"
-import * as routes from "../routes"
-import { useRouter } from "~Generic/hooks/userinterface"
-import { matchesRoute } from "~Generic/lib/routes"
 
 const CreateMainnetAccount = () => (
   <React.Suspense fallback={null}>
@@ -30,55 +26,6 @@ const CreateTestnetAccount = () => (
     <AccountPage accountCreation="testnet" />
   </React.Suspense>
 )
-
-function matchesLastOpenedAccount(account: Account, lastOpenedAccount: Platform.LastOpenedAccount | undefined) {
-  return (
-    lastOpenedAccount?.id === account.id &&
-    lastOpenedAccount.publicKey === account.publicKey &&
-    lastOpenedAccount.testnet === account.testnet
-  )
-}
-
-function LastOpenedAccountManager() {
-  const { accounts, initialized: accountsInitialized } = React.useContext(AccountsContext)
-  const { initialized: settingsInitialized, lastOpenedAccount, setSetting } = React.useContext(SettingsContext)
-  const router = useRouter()
-  const startupResolved = React.useRef(false)
-
-  React.useEffect(() => {
-    if (!accountsInitialized || !settingsInitialized) {
-      return
-    }
-
-    if (!startupResolved.current) {
-      startupResolved.current = true
-
-      if (router.location.pathname === routes.allAccounts() && lastOpenedAccount) {
-        const savedAccount = accounts.find(someAccount => matchesLastOpenedAccount(someAccount, lastOpenedAccount))
-
-        if (savedAccount) {
-          router.history.replace(routes.account(savedAccount.id))
-          return
-        }
-
-        setSetting("lastOpenedAccount", undefined)
-        return
-      }
-    }
-
-    const account = accounts.find(someAccount => matchesRoute(router.location.pathname, routes.account(someAccount.id)))
-
-    if (account) {
-      if (!matchesLastOpenedAccount(account, lastOpenedAccount)) {
-        setSetting("lastOpenedAccount", { id: account.id, publicKey: account.publicKey, testnet: account.testnet })
-      }
-    } else if (router.location.pathname === routes.allAccounts() && lastOpenedAccount) {
-      setSetting("lastOpenedAccount", undefined)
-    }
-  }, [accounts, accountsInitialized, lastOpenedAccount, router.history, router.location.pathname, setSetting, settingsInitialized])
-
-  return null
-}
 
 function Stage2() {
   React.useEffect(() => {
